Forgot Password Functionality
Added the forgot password functionality which is accessible through the login screen.

As described in the issue, the blueprint looks like this:
In login screen --> Blue text that says "Forgot Password?" --> Goes to another screen where it asks for email --> Sends the email once the user presses done --> A modal pops up saying "The email has been sent" --> Takes them back to the Log in screen once they press ok.